+++ dyndnsc/cli.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-Example CLI program
+Main CLI program
 """
 
 import sys
@@ -12,6 +12,10 @@ from dyndnsc.daemon import daemonize
 
 
 def main():
+    '''
+    The main. Initializes the stack, parses command line arguments, and fires
+    requested logic.
+    '''
     from dyndnsc.plugins.manager import DefaultPluginManager
     plugins = DefaultPluginManager()
     plugins.load_plugins()
@@ -20,26 +24,22 @@ def main():
     import argparse
     parser = argparse.ArgumentParser()
 
+    # add the updater protocol options to the CLI:
+    from dyndnsc.updater.manager import updater_classes
+    for kls in updater_classes():
+        kls.register_arguments(parser)
+
+    # add the plugin options to the CLI:
     plugins.options(parser, environ)
 
+    # add generic client options to the CLI:
     parser.add_argument("-d", "--daemon", dest="daemon",
                         help="go into daemon mode (implies --loop)",
                         action="store_true", default=False)
     parser.add_argument("--debug", dest="debug",
                         help="increase logging level to DEBUG",
                         action="store_true", default=False)
-    parser.add_argument("--hostname", dest="hostname",
-                        help="hostname to update", default=None)
-    parser.add_argument("--key", dest="key",
-                        help="your authentication key", default=None)
-    parser.add_argument("--userid", dest="userid",
-                        help="your userid", default=None)
-    parser.add_argument("--password", dest="password",
-                        help="your password", default=None)
-    parser.add_argument("--protocol", dest="protocol",
-                        help="protocol to use for updating IP (default dyndns)",
-                        default='dyndns')
-    parser.add_argument("--method", dest="method",
+    parser.add_argument("--detector", "--method", dest="detector",
                         help="method for detecting your IP (default webcheck)",
                         default='webcheck')
     parser.add_argument("--loop", dest="loop",
@@ -51,6 +51,7 @@ def main():
     parser.add_argument("--version", dest="version",
                         help="show version and exit",
                         action="store_true", default=False)
+
     args = parser.parse_args()
 
     if args.version:
@@ -68,17 +69,21 @@ def main():
     requests_log = logging.getLogger("requests")
     requests_log.setLevel(logging.WARNING)
 
-    if args.hostname is None:
-        parser.error("Please specify a hostname using --hostname")
+    updaters = []
+    for kls in updater_classes():
+        if getattr(args, 'updater_%s' % kls.configuration_key(), False):
+            logging.debug("Gathering initargs for '%s'", kls.configuration_key())
+            initargs = {}
+            for arg_name in kls.init_argnames():
+                val = getattr(args, 'updater_%s_%s' % (kls.configuration_key(), arg_name))
+                if val is not None:
+                    initargs[arg_name] = val
+            updaters.append(kls(**initargs))
 
     config = {}
-    config['hostname'] = args.hostname
-    config['key'] = args.key
-    config['userid'] = args.userid
-    config['password'] = args.password
-    config['protocol'] = args.protocol
-    config['method'] = args.method
+    config['detector'] = args.detector
     config['sleeptime'] = int(args.sleeptime)
+    config['updaters'] = updaters
 
     plugins.configure(args)
     plugins.initialize()