+++ dyndnsc/detector/base.py
@@ -41,7 +41,8 @@ class IPDetector(Subject):
     def set_current_value(self, value):
         self._oldvalue = self.get_current_value()
         self._currentvalue = value
-        log.debug("%s.set_current_value(%s)", self.__class__.__name__, value)
+        if self._oldvalue != value:
+            log.debug("%s.set_current_value(%s)", self.__class__.__name__, value)
         return value
 
     def has_changed(self):