+++ dyndnsc/updater/base.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 
 import logging
-
-import requests
+import textwrap
 
 from ..common.subject import Subject
-from ..common.events import IP_UPDATE_SUCCESS, IP_UPDATE_ERROR
 
 log = logging.getLogger(__name__)
 
@@ -26,51 +24,63 @@ class UpdateProtocol(Subject):
     def updateUrl(self):
         return self.updateurl
 
-    def protocol(self):
-        timeout = 60
-        params = {'myip': self.theip, 'hostname': self.hostname}
-        try:
-            r = requests.get(self.updateUrl(), params=params,
-                             auth=(self.userid, self.password), timeout=timeout)
-        except requests.exceptions.Timeout as exc:
-            log.warning("HTTP timeout(%i) occurred while updating IP at '%s'",
-                      timeout, self.updateUrl(), exc_info=exc)
-            return False
-        else:
-            r.close()
-        log.debug("status %i, %s", r.status_code, r.text)
-        if r.status_code == 200:
-            if r.text.startswith("good "):
-                self.notify_observers(IP_UPDATE_SUCCESS,
-                    "Updated IP address of '%s' to %s"
-                        % (self.hostname, self.theip))
-                return self.theip
-            elif r.text.startswith('nochg'):
-                return self.theip
-            elif r.text == 'nohost':
-                self.notify_observers(IP_UPDATE_ERROR,
-                    "Invalid/non-existant hostname: [%s]" % (self.hostname))
-                return 'nohost'
-            elif r.text == 'abuse':
-                self.notify_observers(IP_UPDATE_ERROR,
-                    "This client is considered to be abusive for hostname '%s'"
-                        % (self.hostname))
-                return 'abuse'
-            elif r.text == '911':
-                self.notify_observers(IP_UPDATE_ERROR, "Service is failing")
-                return '911'
-            elif r.text == 'notfqdn':
-                self.notify_observers(IP_UPDATE_ERROR,
-                    "The provided hostname '%s' is not a valid hostname!"
-                        % (self.hostname))
-                return 'notfqdn'
-            else:
-                self.notify_observers(IP_UPDATE_ERROR,
-                        "Problem updating IP address of '%s' to %s: %s"
-                            % (self.hostname, self.theip, r.text))
-                return r.text
-        else:
-            self.notify_observers(IP_UPDATE_ERROR,
-                "Problem updating IP address of '%s' to %s: %s"
-                    % (self.hostname, self.theip, r.status_code))
-            return 'invalid http status code: %s' % r.status_code
+    def service_url(self):
+        return self.updateUrl()
+
+    @staticmethod
+    def configuration_key():
+        '''
+        This method must be implemented by all updater subclasses. Returns a
+        human readable string identifying the protocol.
+        '''
+        return "none_base_class"
+
+    @classmethod
+    def init_argnames(cls):
+        import inspect
+        return inspect.getargspec(cls.__init__).args[1:]
+
+    @classmethod
+    def _init_argdefaults(cls):
+        import inspect
+        defaults = inspect.getargspec(cls.__init__).defaults
+        if defaults is None:
+            defaults = ()
+        return defaults
+
+    @classmethod
+    def register_arguments(cls, parser):
+        """Register commandline options.
+
+        Implement this method for normal options behavior with protection from
+        OptionConflictErrors. If you override this method and want the default
+        --updater-$name option(s) to be registered, be sure to call super().
+        """
+        cfgkey = cls.configuration_key()
+        parser.add_argument("--updater-%s" % cfgkey,
+                          action="store_true",
+                          dest="updater_%s" % cfgkey,
+                          default=False,
+                          help="Use updater %s: %s" %
+                          (cls.__name__, cls.help()))
+        args = cls.init_argnames()
+        defaults = cls._init_argdefaults()
+        for arg in args[0:len(args) - len(defaults)]:
+            parser.add_argument("--updater-%s-%s" % (cfgkey, arg),
+                              dest="updater_%s_%s" % (cfgkey, arg),
+                              help="")
+        for i, arg in enumerate(args[len(args) - len(defaults):]):
+            parser.add_argument("--updater-%s-%s" % (cfgkey, arg),
+                              dest="updater_%s_%s" % (cfgkey, arg),
+                              default=defaults[i],
+                              help="default: %(default)s")
+
+    @classmethod
+    def help(cls):
+        """Return help for this protocol updater. This will be output as the
+        help section of the --updater-$name option that enables this plugin.
+        """
+        if cls.__doc__:
+            # remove doc section indentation
+            return textwrap.dedent(cls.__doc__)
+        return "(no help available)"