+++ dyndnsc/updater/dnsimple.py
@@ -11,23 +11,25 @@ Since dnsimple.com is a paid service, I have not had a chance to test this yet.
 
 import logging
 
+from .base import UpdateProtocol
+
 from dnsimple_dyndns import DNSimple
 
 log = logging.getLogger(__name__)
 
 
-class UpdateProtocolDnsimple(object):
+class UpdateProtocolDnsimple(UpdateProtocol):
     """Protocol handler for https://dnsimple.com/"""
 
-    def __init__(self, hostname, key, url=None, **kwargs):
+    def __init__(self, hostname, key, service_url=None, **kwargs):
         self.theip = None
         self._recordname, _, self._domain = hostname.partition(".")
         self.handler = DNSimple(domain=self._domain,
                                 domain_token=key)
-        if url is not None:
+        if service_url is not None:
             # The url must be a format string like this:
             #      'https://dnsimple.com/domains/%s/records'
-            self.handler._baseurl = url % self._domain
+            self.handler._baseurl = service_url % self._domain
 
     @staticmethod
     def configuration_key():